LPS-55111 Gradle plugin - move files
