+++ modules/sdk/gradle-plugins-source-formatter/bnd.bnd
@@ -2,4 +2,6 @@ Bundle-Name: Liferay Gradle Plugins Source Formatter
 Bundle-SymbolicName: com.liferay.gradle.plugins.source.formatter
 Bundle-Version: 1.0.0
 Export-Package: com.liferay.gradle.plugins.source.formatter
-Include-Resource: classes
+Include-Resource:\
+	classes,\
+	@lib/com.liferay.source.formatter.jar!/com/liferay/source/formatter/SourceFormatterArgs.class
