+++ modules/sdk/gradle-plugins-source-formatter/samples/settings.gradle
@@ -1,0 +1,1 @@
+include "samples"
