+++ modules/util/source-formatter/bnd.bnd
@@ -6,7 +6,6 @@ Import-Package:\
 	!javax.servlet.http,\
 	!org.apache.maven.plugin,\
 	!org.apache.tools.ant.*,\
-	!org.gradle.api.*,\
 	*
 Include-Resource:\
 	classes,\
