--- modules/util/source-formatter/samples/build.gradle
@@ -1,16 +1,0 @@
-buildscript {
-	repositories {
-		mavenLocal()
-		mavenCentral()
-	}
-
-	dependencies {
-		classpath 'com.liferay:com.liferay.source.formatter:1.0.0'
-	}
-}
-
-apply plugin: 'com.liferay.source.formatter'
-
-sourceFormatter {
-	baseDir="src/main"
-}
