--- modules/util/source-formatter/samples/settings.gradle
@@ -1,1 +1,0 @@
-include 'samples'
