--- modules/util/source-formatter/src/META-INF/gradle-plugins/com.liferay.source.formatter.properties
@@ -1,1 +1,0 @@
-implementation-class=com.liferay.source.formatter.gradle.SourceFormatterPlugin
